Internal Functions

Documentation for project internals. This is useful mainly to devs or individuals wanting to take apart the package.

High level internals

MembraneEOS.ChemicalParameters — Type

Define a set of parameters that describe a chemical. This method uses an optional (positional) name, and additional keyword arguments (with units) described below.

Note

It is highly reccomended to stay within the units commonly used through database lookups to avoid the risk of unit errors.

Parameter	unit

molecular_weight	g/mol

critical_temperature	K

critical_pressure	atm

acentric factor	no units

characteristic_temperature	K

characteristic_pressure	MPa

characteristic_density	g/cm3

Low level internals

MembraneEOS.UnorderedChemicalPair — Type

Structs and globals

MembraneEOS.create_kij_dict_from_csv! — Function

KIJ database reading functions and lookups