EOS Parameters

Parameters can be specified directly or looked up.

CubicParameters

MembraneEOS.CubicParameters — Type

CubicParameters(s::String)

Attempt to look up a set of cubic parameters, returns missing if not found.

CubicParameters(critical_temperature, critical_pressure, acentric_factor, [molecular_weight])

Directly create some chemical's CubicParameters. Molecular weight is optional as it isn't necessary for standard functionality.

Parameters	Units

Critical Temperature	K

Critical Pressure	atm

Acentric Factor	n/a

Molecular Weight	g/mol

SanchezLacombeParameters

MembraneEOS.SanchezLacombeParameters — Type

SanchezLacombeParameters(s::String)

Attempt to look up a set of Sanchez Lacombe parameters, returns missing if not found.

SanchezLacombeParameters(characteristic_pressure, characteristic_temperature, characteristic_density, molecular_weight)

Directly create some chemical's SanchezLacombeParameters.

Parameters	Units

Characteristic Temperature	K

Characteristic Pressure	MPa

Characteristic Density	g/cm3

Molecular Weight	g/mol